+++ src/TwigComponent/src/ComponentRenderer.php
@@ -14,7 +14,8 @@ namespace Symfony\UX\TwigComponent;
 use Symfony\Component\PropertyAccess\PropertyAccessorInterface;
 use Symfony\Contracts\EventDispatcher\EventDispatcherInterface;
 use Symfony\UX\TwigComponent\Attribute\ExposeInTemplate;
-use Symfony\UX\TwigComponent\EventListener\PreRenderEvent;
+use Symfony\UX\TwigComponent\Event\PreCreateForRenderEvent;
+use Symfony\UX\TwigComponent\Event\PreRenderEvent;
 use Twig\Environment;
 use Twig\Extension\EscaperExtension;
 
@@ -40,6 +41,14 @@ final class ComponentRenderer implements ComponentRendererInterface
 
     public function createAndRender(string $name, array $props = []): string
     {
+        $event = new PreCreateForRenderEvent($name, $props);
+        $this->dispatcher->dispatch($event);
+
+        // allow the process to be short-circuited
+        if (null !== $rendered = $event->getRenderedString()) {
+            return $rendered;
+        }
+
         return $this->render($this->factory->create($name, $props));
     }